Adds some code fixes
