+++ Controller/Customer.php
@@ -1,0 +1,63 @@
+<?php
+declare(strict_types=1);
+/**
+ * Magetarian_CustomerTwoFactorAuth
+ *
+ * PHP version 7.x
+ *
+ * @category PHP
+ * @package  Magetarian\CustomerTwoFactorAuth
+ */
+
+namespace Magetarian\CustomerTwoFactorAuth\Controller;
+
+use Magento\Customer\Model\Session;
+use Magento\Framework\App\Action\Action;
+use Magento\Framework\App\Action\Context;
+use Magento\Framework\App\RequestInterface;
+use Magento\Framework\App\ResponseInterface;
+use Magento\Framework\Exception\NotFoundException;
+
+/**
+ * Class for Customer Implementation
+ *
+ * @package Magetarian\CustomerTwoFactorAuth
+ */
+abstract class Customer extends Action
+{
+    /**
+     * @var Session
+     */
+    protected $customerSession;
+
+
+    /**
+     * Configuration constructor.
+     *
+     * @param Context $context
+     * @param Session $customerSession
+     */
+    public function __construct(
+        Context $context,
+        Session $customerSession
+    ) {
+        parent::__construct($context);
+        $this->customerSession = $customerSession;
+    }
+
+
+    /**
+     * Dispatch request
+     *
+     * @param RequestInterface $request
+     * @return ResponseInterface
+     * @throws NotFoundException
+     */
+    public function dispatch(RequestInterface $request)
+    {
+        if (!$this->customerSession->authenticate()) {
+            $this->_actionFlag->set('', 'no-dispatch', true);
+        }
+        return parent::dispatch($request);
+    }
+}
