+++ etc/frontend/routes.xml
@@ -6,7 +6,7 @@
   -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:App/etc/routes.xsd">
     <router id="standard">
-        <route id="twoFactorAuth" frontName="twoFactorAuth">
+        <route id="twofactorauth" frontName="twofactorauth">
             <module name="Magetarian_CustomerTwoFactorAuth" />
         </route>
     </router>
