+++ view/frontend/templates/customer_account/two_factor_auth_config.phtml
@@ -47,5 +47,5 @@ $viewModel = $block->getViewModel();
         </div>
     </div>
 <?php else:?>
-    <?= $block->escapeHtml(__('There are no two factor authroization providers configured.'));?>
+    <?= $block->escapeHtml(__('There are no two factor authorization providers configured.'));?>
 <?php endif; ?>
